refactor: organize button component into its own directory
