+++ src/components/article/ArticleActions.tsx
@@ -1,6 +1,6 @@
 import Link from 'next/link';
 import Image from 'next/image';
-import { DeleteArticleButton } from '@/components/article';
+import { DeleteArticleButton, EditArticleButton } from '@/components/article';
 
 export default function ArticleActions({
     slug,
@@ -38,10 +38,8 @@ export default function ArticleActions({
                       <i className="ion-heart"></i>
                       &nbsp; Favorite Article <span className="counter">(29)</span>
                   </button>
-                <button className="btn btn-sm btn-outline-secondary">
-                    <i className="ion-edit"></i> Edit Article
-                </button>
                 */}
+                <EditArticleButton slug={slug} />
                 <DeleteArticleButton slug={slug} />
             </div>
         </div>
